Add data types for mittens journals
There's some missing functionality in the time package, which I patched &
pull-requested (https://github.com/haskell/time/pull/8). While I'm waiting for
that to be added to the official time package, I created some orphan instances
myself, which are in Mittens.Time.

The rest of the stuff is to add journal-keeping functionality to mittens. The
functionality isn't there yet, but the data types exist.