Test - encode/decode a variable number of times.
There is a test, which tests the Aeson transparency. Essentially, it takes a
Journal, converts it into JSON, and then reads it back from JSON. It then makes
sure the two objects are identical (It doesn't quite work that way, but
whatever). This test I just wrote will do that a variable number of times, and
make sure (ideally) that it will work any number of times (n >= 1).